requirements.txt
google-cloud-storage>=2.17.0
google-cloud-aiplatform>=1.66.0   # for Vertex AI / genai
google-adk>=0.2.0                 # ADK SDK (adjust if version differs in your env)
cairosvg>=2.7.0
pillow>=10.0.0
ipython>=8.0.0
python-dotenv>=1.0.0

